# registration-application-backend/config/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Auth API",
        "version": "1.0.0",
        "description": "API for user registration and login with JWT authentication"
    },
    "servers": [
        {
            "url": "http://localhost:3000/api"
        }
    ],
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "paths": {
        "/auth/register": {
            "post": {
                "summary": "Register a new user",
                "tags": ["Auth"],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "username": { "type": "string" },
                                    "email": { "type": "string" },
                                    "password": { "type": "string" }
                                },
                                "required": ["username", "email", "password"]
                            }
                        }
                    }
                },
                "responses": {
                    "201": { "description": "User registered successfully" },
                    "400": { "description": "Bad request" },
                    "409": { "description": "User already exists" }
                }
            }
        },
        "/auth/login": {
            "post": {
                "summary": "Login user",
                "tags": ["Auth"],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "email": { "type": "string" },
                                    "password": { "type": "string" }
                                },
                                "required": ["email", "password"]
                            }
                        }
                    }
                },
                "responses": {
                    "200": { "description": "Login successful, returns JWT" },
                    "401": { "description": "Invalid credentials" }
                }
            }
        },
        "/auth/protected": {
            "get": {
                "summary": "Protected route requiring JWT",
                "tags": ["Auth"],
                "security": [
                    { "BearerAuth": [] }
                ],
                "responses": {
                    "200": {
                        "description": "Access granted to protected route",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": { "type": "string" },
                                        "user": { "type": "object" }
                                    }
                                }
                            }
                        }
                    },
                    "401": { "description": "Unauthorized" }
                }
            }
        },
        "/auth/forgot-password": {
            "get": {
                "summary": "Protected route requiring JWT",
                "tags": ["Auth"],
                "security": [
                    { "BearerAuth": [] }
                ],
                "responses": {
                    "200": {
                        "description": "Access granted to protected route",
                        "content": {
                            "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "email": { "type": "string" }
                                },
                                "required": ["email"]
                            }
                        }
                        }
                    },
                    "401": { "description": "Unauthorized" }
                }
            }
        },
        "/auth/reset-password": {
            "get": {
                "summary": "Reset password API",
                "tags": ["Auth"],
                "security": [
                    { "BearerAuth": [] }
                ],
                "responses": {
                    "200": {
                        "description": "Access granted to protected route",
                        "content": {
                            "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "newPassword": { "type": "string" },
                                    "confirmPassword": { "type": "string" }
                                },
                                "required": ["newPassword", "confirmPassword"]
                            }
                        }
                        }
                    },
                    "401": { "description": "Unauthorized" }
                }
            }
        }
    }
}